Rename "message" types uniformly to `FooMessage`
Some message types (e.g., `BaseMessage` and `AuthMessage`) used the
full spelled-out word "Message" in their names, while others (like
`ChanMsg` and `stateChangedMsg`) used "Msg". For consistency, rename
the latter to be more consistent with the former:

* `ChanMsg` → `ChanMessage`
* `stateChangedMsg` → `stateChangedMessage`
