+++ eventListener.go
@@ -158,7 +158,7 @@ func (l *EventListener) maybeCall(app *App, eventData EventData) {
 }
 
 /* Functions */
-func (app *App) callEventListeners(eventType string, msg websocket.ChanMsg) {
+func (app *App) callEventListeners(eventType string, msg websocket.ChanMessage) {
 	listeners, ok := app.eventListeners[eventType]
 	if !ok {
 		// no listeners registered for this event type
